Kernel: Introduce InterruptSupervisor class
The idea behind this is that PCI devices could be responsible for
many interrupt lines at once (IRQs and/or MSIs).
For this purpose all PCI::Device are inherently an InterruptSupervisor.
Upper classes can use this functionality to spawn multiple
IRQHandlers for example.

One usage of this change is within the refactoring of the disk subsystem
where I want to have an IDEController class that supervises
2 PATAChannels.
Later, when we have MSI support, the code for NVMe support could use
this change again.

A little bit about lifetime cycle - this object has one reference to it
in the InterruptManagement object in a Vector.
It could be deleted when power event occurs (e.g. reboot, shutdown) or
even hotplug if we happen to support such functionality.
When that happens, all registered interrupt handlers should be
deleted too, hence keeping the system clean from those handlers.