Kernel+Userland: Rename Resolution => ModeSetting in most places
In addition the name being changed, the concept of setting a resolution
is reduced to setting a resolution of a framebuffer console, anything
else in the graphics subsystem use the term ModeSetting, which represent
the fact that bare metal hardware can't just use arbitrary resolution
with arbitrary width and height, but specific values to determine the
correct display timings.

There are still some leftovers in the WindowServer code, a remainder
that we can't (yet) change resolutions on bare metal hardware. This
commit is another step into allowing that to happen.