Kernel/Graphics: Simplify EDID initialization and obtainment
This is done in two aspects:
1. We make the get_edid in the DisplayConnector as non virtual method so
it must be implemented only once. This enforces consistency across the
subsystem.
2. Because get_edid in the DisplayConnector class is non-virtual method,
now each driver must initialize the EDID after constructing a
DisplayConnector object. Thus, we make all instances to use m_edid_bytes
for later usage such as obtainment of the EDID from userspace.
Because all related EDID code in the kernel relies on the m_edid_bytes
to be valid, we also ensure all drivers are able to supply an EDID. For
paravirtualized hardware, such as the VMWareSVGA and VirtualBox graphics
we generate a general fake EDID so userspace can still utilize these
devices.