Kernel/Graphics: Re-organize the Bochs driver
The parent class in the inheritance hierarchy is BochsDisplayConnector
for this driver, and this class helps to drive the original bochs VGA
device. Then, there's the QEMUDisplayConnector class that helps to drive
the QEMU std-vga (also the bochs-display) device.
VirtualBox users could utilize for now the BochsDisplayConnector
capabilities until we have better support for that device.