Kernel/PCI: Re-design the entire subsystem
This commit puts the PCI bus subsystem inline with the standard that was
set by the USB bus subsystem.

This commit removes the responsibility to detect devices from all
management singletons which in turn makes the whole device detection
sequence to be asynchronous from the main init sequence.

This is done by removing the PCI::DeviceIdentifier as passing object
to kernel drivers. Instead, it's expected that each PCI::HostController
object enumerates all attaches busses and create PCI::Bus objects.

Then it should find the appropriate device driver, similarly to how it
is done in the USB bus subsystem - each device driver is registered in
a list which can iterated.

A PCI device driver has a basic functionality of probing for appropriate
devices, attaching and detaching as desired.

To ensure that we can enumerate all PCI devices and then find their
appropriate device driver, we should make PCI::Device as standalone,
atomic ref-counted object. All PCI drivers and their related structures
should then hold a reference to a used PCI::Device, similarly to how it
is done in the USB bus subsystem.

To optimize attaching devices to drivers, each PCI::Driver advertises
its responsible ClassID. The PCI::Access code then registers the actual
PCI::Driver into the appropriate list based on the numeric value of that
property.
It does the same when registering a PCI::Device - it registers it into
the appropriate list based on that value as well.

When registering a driver, the PCI::Access code searches for matching
devices in the appropriate list and asks the PCI::Driver to probe them
accordingly.
However, when registering a device, we try to find the best PCI::Driver
that is currently registered, based not only a mere PCI::HardwareID
match, but it checks for specific details such as Revision ID, Prog-IF
and SubsystemID of the device against a match from the PCI::Driver.

Because the subsystem will be able to find the devices asynchronously,
the UNMAP_AFTER_INIT tag was removed from initialization paths on many
PCI drivers.

To acommodate flexible sequence of inserting devices, we also remove the
the AudioManagement singleton.
The reason for removing this class was because it is essentially no
longer playing any major role in booting the system, as it is not
calling the appropriate drivers anymore.

We add a basic set of PCI drivers, so the system can still boot in QEMU
with graphics, networking and storage.